+++ packages/dnb-ui-lib/src/components/number/Number.js
@@ -12,7 +12,8 @@ import {
   isTrue,
   validateDOMAttributes,
   registerElement,
-  extend
+  extend,
+  PLATFORM_MAC
 } from '../../shared/component-helper'
 import { createSpacingClasses } from '../space/SpacingHelper'
 
@@ -282,23 +283,26 @@ export const format = (
     )
 
     display = formatNumber(cleanedNumber, locale, opts)
-
-    // change the position of minus if it's first
-    const first = display[0]
-    if (first === '−' || first === '-') {
-      display = display.replace(/^(−|-)([^\d]+)(.*)/g, '$2$1$3')
-    }
+    display = cleanupMinus(display)
 
     // aria options
     opts.currencyDisplay = 'name'
     aria = formatNumber(cleanedNumber, locale, opts)
+    aria = enhanceSR(cleanedNumber, aria, locale)
+
+    // get only the currency name
+    // const num = aria.replace(/([^0-9])+$/g, '')
+    // const name = aria.replace(num, '')
+    // aria = cleanedNumber + name
   } else {
-    display = formatNumber(parseFloat(value), locale, opts)
+    display = formatNumber(value, locale, opts)
+    display = cleanupMinus(display)
 
     // fix for NDVA to make sure we read the number, we add a minium fraction digit (decimal)
     if (typeof opts.minimumFractionDigits === 'undefined') {
-      opts.minimumFractionDigits = 1
-      aria = formatNumber(parseFloat(value), locale, opts)
+      opts.minimumFractionDigits = 1 // NVDA fix
+      aria = formatNumber(value, locale, opts)
+      aria = enhanceSR(value, aria, locale)
     }
   }
 
@@ -310,9 +314,53 @@ export const format = (
   return returnAria ? { number: display, aria, locale } : display
 }
 
+const cleanupMinus = display => {
+  // change the position of minus if it's first
+  // check for two minus - −
+  // check also for hyphen ‐
+  // check also for dashes ‒  –  —  ―
+
+  const reg = '^(-|−|‐|‒|–|—|―)'
+
+  // check for first and second char
+  const first = display[0]
+  const second = display[1]
+
+  if (new RegExp(reg).test(first)) {
+    // if second is number
+    if (parseFloat(second) > 0) {
+      // then do not swap
+      display = display.replace(new RegExp(reg + '(.*)'), '-$2')
+    } else {
+      // then first has to be currency
+      display = display.replace(new RegExp(reg + '([^\\d]+)(.*)'), '$2-$3')
+    }
+  }
+
+  return display
+}
+
+const enhanceSR = (value, aria) => {
+  // Enhance VO support on mobile devices
+  // Numbers under 99.999 are read out correctly, but only if we remove the spaces
+  // Potential we could also check for locale: && /no|nb|nn/.test(locale)
+  // but leave it for now without this ectra check
+  if (
+    typeof navigator !== 'undefined' &&
+    navigator.platform.match(new RegExp(PLATFORM_MAC)) !== null &&
+    Math.abs(parseFloat(value)) <= 99999
+  ) {
+    aria = String(aria).replace(/\s([0-9])/g, '$1')
+  }
+
+  aria = cleanupMinus(aria)
+
+  return aria
+}
+
 export const formatNumber = (number, locale, options = {}) => {
-  if (typeof number.toLocaleString === 'function') {
-    return number.toLocaleString(locale, options)
+  if (typeof Number.toLocaleString === 'function') {
+    return parseFloat(number).toLocaleString(locale, options)
   } else if (
     typeof Intl !== 'undefined' &&
     typeof Intl.NumberFormat === 'function'