+++ packages/dnb-ui-lib/stories/helperStories.js
@@ -6,13 +6,29 @@
 import React from 'react'
 import { Wrapper, Box } from './helpers'
 import styled from '@emotion/styled'
-import { Button } from 'dnb-ui-lib/src'
+import { Button, Section } from 'dnb-ui-lib/src'
 
 const stories = []
 export default stories
 
 const SkipLinkReset = styled.div`
   /* empty */
+`
+const WidthWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+`
+const WidthWrapperInner = styled.div`
+  /* for large screesn */
+  width: 84vw; /* apr. the same as 60rem */
+  max-width: 60rem;
+  padding: 0 1rem;
+
+  /* for small screesn */
+  @media (max-width: 40em) {
+    width: 95vw;
+    padding: 0;
+  }
 `
 
 stories.push([
@@ -38,11 +54,17 @@ stories.push([
   'Section',
   () => (
     <Wrapper>
-      <Box>
-        <div className="dnb-section dnb-section--divider">
-          <Button text="Next" icon="chevron_right" />
-        </div>
-      </Box>
+      <WidthWrapper>
+        <WidthWrapperInner>
+          <Section style_type="emerald-green" spacing>
+            <p className="dnb-p">
+              Tempor auctor mi vel sociosqu suspendisse eros fames congue
+              at
+            </p>
+            <Button text="Next" icon="chevron_right" />
+          </Section>
+        </WidthWrapperInner>
+      </WidthWrapper>
     </Wrapper>
   )
 ])